+++ lib/ruby/openai/client.rb
@@ -33,7 +33,32 @@ module OpenAI
       @files ||= OpenAI::Files.new(access_token: @access_token)
     end
 
-    def search(engine:, query:, documents: nil, file: nil, version: default_version)
+    # rubocop:disable Layout/LineLength
+    # rubocop:disable Metrics/ParameterLists
+    def search(engine:, query: nil, documents: nil, file: nil, version: default_version, parameters: {})
+      return legacy_search(engine: engine, query: query, documents: documents, file: file, version: version) if query || documents || file
+
+      self.class.post(
+        "/#{version}/engines/#{engine}/search",
+        headers: {
+          "Content-Type" => "application/json",
+          "Authorization" => "Bearer #{@access_token}"
+        },
+        body: parameters.to_json
+      )
+    end
+    # rubocop:enable Layout/LineLength
+    # rubocop:enable Metrics/ParameterLists
+
+    private
+
+    # rubocop:disable Metrics/MethodLength
+    # rubocop:disable Layout/LineLength
+    def legacy_search(engine:, query:, documents: nil, file: nil, version: default_version)
+      warn "[DEPRECATION] Passing `query`, `documents` or `file` directly to `Client#search` is deprecated and will be removed in a future version of ruby-openai.
+        Please nest these terms within `parameters` instead, like this:
+        client.search(engine: 'davinci', parameters: { query: 'president', documents: %w[washington hospital school] })
+      "
       self.class.post(
         "/#{version}/engines/#{engine}/search",
         headers: {
@@ -43,8 +68,8 @@ module OpenAI
         body: { query: query }.merge(documents_or_file(documents: documents, file: file)).to_json
       )
     end
-
-    private
+    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Layout/LineLength
 
     def default_version
       "v1".freeze