+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - 2021-04-18
+
+### Changed
+
+- Add Client#search(parameters:) to allow passing `max_rerank` or `return_metadata`.
+- Deprecate Client#search with query, file or document parameters at the top level.
+- Thanks [@stevegeek](https://github.com/stevegeek) for pointing this issue out!
+
 ## [1.2.1] - 2021-04-11
 
 ### Added